[REFACTOR] navigation interface분리(#139)
