+++ LionHeart-iOS/LionHeart-iOS/Global/Enums/TokenState.swift
@@ -1,0 +1,13 @@
+//
+//  TokenState.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+enum TokenState {
+    case valid
+    case expired
+}
