+++ LionHeart-iOS/LionHeart-iOS/Global/Enums/UserState.swift
@@ -1,0 +1,13 @@
+//
+//  UserState.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+enum UserState {
+    case verified
+    case nonVerified
+}
