+++ LionHeart-iOS/LionHeart-iOS/Scenes/Curriculum/ViewControllers/CurriculumViewController.swift
@@ -11,9 +11,7 @@ import UIKit
 import SnapKit
 import Lottie
 
-protocol CurriculumNavigation: BarNavigation, ExpireNavigation {
-    func articleListCellTapped(itemIndex: Int)
-}
+
 
 protocol CurriculumManager {
     func getCurriculumServiceInfo() async throws -> UserInfoData
