+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/ArticleDetailModalNavigation.swift
@@ -1,0 +1,10 @@
+//
+//  ArticleDetailModalNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol ArticleDetailModalNavigation: ExpireNavigation, DismissNavigation {}
