+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/ChallengeNavigation.swift
@@ -1,0 +1,10 @@
+//
+//  ChallengeNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol ChallengeNavigation: BarNavigation {}
