+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/CompleteOnbardingNavigation.swift
@@ -1,0 +1,12 @@
+//
+//  CompleteOnbardingNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol CompleteOnbardingNavigation: ExpireNavigation {
+    func startButtonTapped()
+}
