+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/CurriculumNavigation.swift
@@ -1,0 +1,12 @@
+//
+//  CurriculumNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol CurriculumNavigation: BarNavigation, ExpireNavigation {
+    func articleListCellTapped(itemIndex: Int)
+}
