+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/LoginNavigation.swift
@@ -1,0 +1,12 @@
+//
+//  LoginNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol LoginNavigation: AnyObject {
+    func checkUserIsVerified(userState: UserState, kakaoToken: String?)
+}
