+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/MyPageNavigation.swift
@@ -1,0 +1,10 @@
+//
+//  MyPageNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol MyPageNavigation: ExpireNavigation, PopNavigation {}
