+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/OnboardingNavigation.swift
@@ -1,0 +1,12 @@
+//
+//  OnboardingNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol OnboardingNavigation: ExpireNavigation, PopNavigation {
+    func onboardingCompleted(data: UserOnboardingModel)
+}
