+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/SplashNavigation.swift
@@ -1,0 +1,12 @@
+//
+//  SplashNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol SplashNavigation: AnyObject {
+    func checkToken(state: TokenState)
+}
