+++ LionHeart-iOS/LionHeart-iOS/Scenes/NavigationInterface/TodayNavigation.swift
@@ -1,0 +1,12 @@
+//
+//  TodayNavigation.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol TodayNavigation: ExpireNavigation, BarNavigation {
+    func todayArticleTapped(articleID: Int)
+}
